+++ autogen/agentchat/realtime_agent/__init__.py
@@ -6,5 +6,13 @@ from .audio_adapters import TwilioAudioAdapter, WebSocketAudioAdapter
 from .function_observer import FunctionObserver
 from .realtime_agent import RealtimeAgent
 from .realtime_observer import RealtimeObserver
+from .realtime_swarm import SwarmableRealtimeAgent
 
-__all__ = ["FunctionObserver", "RealtimeAgent", "RealtimeObserver", "TwilioAudioAdapter", "WebSocketAudioAdapter"]
+__all__ = [
+    "FunctionObserver",
+    "RealtimeAgent",
+    "RealtimeObserver",
+    "SwarmableRealtimeAgent",
+    "TwilioAudioAdapter",
+    "WebSocketAudioAdapter",
+]
