+++ autogen/agentchat/realtime_agent/realtime_agent.py
@@ -5,14 +5,11 @@
 from logging import Logger, getLogger
 from typing import Any, Callable, Optional, TypeVar, Union
 
-import anyio
-from asyncer import create_task_group, syncify
+from asyncer import create_task_group
 
-from ... import SwarmAgent
 from ...tools import Tool
 from ..agent import Agent
 from ..conversable_agent import ConversableAgent
-from .clients import Role
 from .clients.realtime_client import RealtimeClientProtocol, get_client
 from .function_observer import FunctionObserver
 from .realtime_observer import RealtimeObserver
@@ -20,20 +17,6 @@ from .realtime_observer import RealtimeObserver
 F = TypeVar("F", bound=Callable[..., Any])
 
 global_logger = getLogger(__name__)
-
-SWARM_SYSTEM_MESSAGE = (
-    "You are a helpful voice assistant. Your task is to listen to user and to coordinate the tasks based on his/her inputs."
-    "Only call the 'answer_task_question' function when you have the answer from the user."
-    "You can communicate and will communicate using audio output only."
-)
-
-QUESTION_ROLE: Role = "user"
-QUESTION_MESSAGE = (
-    "I have a question/information for myself. DO NOT ANSWER YOURSELF, GET THE ANSWER FROM ME. "
-    "repeat the question to me **WITH AUDIO OUTPUT** and then call 'answer_task_question' AFTER YOU GET THE ANSWER FROM ME\n\n"
-    "The question is: '{}'\n\n"
-)
-QUESTION_TIMEOUT_SECONDS = 20
 
 
 class RealtimeAgent(ConversableAgent):
@@ -97,12 +80,6 @@ class RealtimeAgent(ConversableAgent):
             [Agent, None], RealtimeAgent.check_termination_and_human_reply, remove_other_reply_funcs=True
         )
 
-        self._answer_event: anyio.Event = anyio.Event()
-        self._answer: str = ""
-        self._start_swarm_chat = False
-        self._initial_agent: Optional[SwarmAgent] = None
-        self._agents: Optional[list[SwarmAgent]] = None
-
     @property
     def logger(self) -> Logger:
         """Get the logger for the agent."""
@@ -126,37 +103,13 @@ class RealtimeAgent(ConversableAgent):
         """
         self._observers.append(observer)
 
-    def register_swarm(
-        self,
-        *,
-        initial_agent: SwarmAgent,
-        agents: list[SwarmAgent],
-        system_message: Optional[str] = None,
-    ) -> None:
-        """Register a swarm of agents with the Realtime Agent.
+    async def start_observers(self) -> None:
+        for observer in self._observers:
+            self._tg.soonify(observer.run)(self)
 
-        Args:
-            initial_agent (SwarmAgent): The initial agent.
-            agents (list[SwarmAgent]): The agents in the swarm.
-            system_message (str): The system message for the agent.
-        """
-        logger = self.logger
-        if not system_message:
-            if self.system_message != "You are a helpful AI Assistant.":
-                logger.warning(
-                    "Overriding system message set up in `__init__`, please use `system_message` parameter of the `register_swarm` function instead."
-                )
-            system_message = SWARM_SYSTEM_MESSAGE
-
-        self._oai_system_message = [{"content": system_message, "role": "system"}]
-
-        self._start_swarm_chat = True
-        self._initial_agent = initial_agent
-        self._agents = agents
-
-        self.register_realtime_function(name="answer_task_question", description="Answer question from the task")(
-            self.set_answer
-        )
+        # wait for the observers to be ready
+        for observer in self._observers:
+            await observer.wait_for_ready()
 
     async def run(self) -> None:
         """Run the agent."""
@@ -164,13 +117,8 @@ class RealtimeAgent(ConversableAgent):
         async with create_task_group() as self._tg:
             # connect with the client first (establishes a connection and initializes a session)
             async with self._realtime_client.connect():
-                # start the observers
-                for observer in self._observers:
-                    self._tg.soonify(observer.run)(self)
-
-                # wait for the observers to be ready
-                for observer in self._observers:
-                    await observer.wait_for_ready()
+                # start the observers and wait for them to be ready
+                await self.start_observers()
 
                 # iterate over the events
                 async for event in self.realtime_client.read_events():
@@ -209,71 +157,3 @@ class RealtimeAgent(ConversableAgent):
             return tool
 
         return _decorator
-
-    def reset_answer(self) -> None:
-        """Reset the answer event."""
-        self._answer_event = anyio.Event()
-
-    def set_answer(self, answer: str) -> str:
-        """Set the answer to the question."""
-        self._answer = answer
-        self._answer_event.set()
-        return "Answer set successfully."
-
-    async def get_answer(self) -> str:
-        """Get the answer to the question."""
-        await self._answer_event.wait()
-        return self._answer
-
-    async def ask_question(self, question: str, question_timeout: int) -> None:
-        """Send a question for the user to the agent and wait for the answer.
-        If the answer is not received within the timeout, the question is repeated.
-
-        Args:
-            question: The question to ask the user.
-            question_timeout: The time in seconds to wait for the answer.
-        """
-        self.reset_answer()
-        await self._realtime_client.send_text(role=QUESTION_ROLE, text=question)
-
-        async def _check_event_set(timeout: int = question_timeout) -> bool:
-            for _ in range(timeout):
-                if self._answer_event.is_set():
-                    return True
-                await anyio.sleep(1)
-            return False
-
-        while not await _check_event_set():
-            await self._realtime_client.send_text(role=QUESTION_ROLE, text=question)
-
-    def check_termination_and_human_reply(
-        self,
-        messages: Optional[list[dict[str, Any]]] = None,
-        sender: Optional[Agent] = None,
-        config: Optional[Any] = None,
-    ) -> tuple[bool, Union[str, None]]:
-        """Check if the conversation should be terminated and if the agent should reply.
-
-        Called when its agents turn in the chat conversation.
-
-        Args:
-            messages: list of dict
-                the messages in the conversation
-            sender: Agent
-                the agent sending the message
-            config: any
-                the config for the agent
-        """
-        if not messages:
-            return False, None
-
-        async def get_input() -> None:
-            async with create_task_group() as tg:
-                tg.soonify(self.ask_question)(
-                    QUESTION_MESSAGE.format(messages[-1]["content"]),
-                    question_timeout=QUESTION_TIMEOUT_SECONDS,
-                )
-
-        syncify(get_input)()
-
-        return True, {"role": "user", "content": self._answer}  # type: ignore[return-value]
